+++ module/apps/compendium-browser.js
@@ -287,7 +287,7 @@ export class CompendiumBrowser extends Application {
       // Add CR filters
       if (item.data.type === "npc") {
         const cr = getProperty(item.data, "data.details.cr");
-        if (cr && !this.extraFilters["data.details.cr"].includes(cr)) this.extraFilters["data.details.cr"].push(cr);
+        if (cr && !this.extraFilters["data.details.cr"].includes(cr)) this.extraFilters["data.details.cr"].push(parseFloat(cr));
       }
     }
 
@@ -509,7 +509,9 @@ export class CompendiumBrowser extends Application {
       {
         path: "data.details.cr",
         label: "CR",
-        items: this.extraFilters["data.details.cr"].sort().reduce((cur, o) => {
+        items: this.extraFilters["data.details.cr"].sort(function(a, b) {
+          return a - b;
+        }).reduce((cur, o) => {
           cur.push({ key: o, name: CR.fromNumber(o) });
           return cur;
         }, []),